Split off hash.h from util.h
